Added Robust LMM package and method
The idea is that I am able to generate a robust model for each dependent variable. 

Assumption: I am assuming that if the RMEL results for the robust linear model are similar to the REML results for the non-robust model then I can be confident of the results. (Note: This may change as my understanding of Robustness changes).