Data Functions
Added a function that generates different functions to be analyzed. Th function name is :

model_data_Target_Session

Test this function in the Model script. Also, fitted models with REML because when they are compared using anova, that function re-fits the models using ML.

Added a package called robustlmm. The purpose of this package is to compare the model fit using a robust fir method to a model fit using the another method. This would primarily be done when some assumptions are violated. The hope is that this will reveal if the violations matters.