Census app
This is the a partially complete sense app based on the shiny tutorial in Rstudio. It displays a percentage map of the populations across the country. 

The app is NOT dynamic (even though there are buttons to change the display). I stopped at this point because I felt as though I already discovered all that I needed.